fixes to delayManager; freeDrinks has no pre-onset if no delaymanager; small fix to compile
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@2394 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ classes/util/trialRecordProcessing/extractFieldAndEnsure.m
@@ -248,6 +248,6 @@ end
 function out = getCorrect(trialDetails)
 out=nan;
 if isfield(trialDetails,'correct') && ~isempty(trialDetails.correct)
-    out=trialDetails.correct;
+    out=double(trialDetails.correct);
 end
 end
